+++ app/[locale]/projects/[projectId]/collaboration/invite/actions/addCollaboratorAction.ts
@@ -1,7 +1,7 @@
 "use server";
 
 import { executeGraphQLMutation } from "@/utils/server/graphqlServerActionHandler";
-import logger from "@/utils/server/logger";
+import logger, {prepareObjectForLogs} from "@/utils/server/logger";
 import { CollaboratorResponse } from "@/app/types";
 import { AddProjectCollaboratorDocument } from "@/generated/graphql";
 
@@ -15,6 +15,8 @@ export async function addProjectCollaboratorAction({
   accessLevel: string;
 }): Promise<CollaboratorResponse> {
   try {
+    logger.debug(await prepareObjectForLogs({ projectId, email, accessLevel }), "Adding project collaborator");
+
     // Execute the mutation using the shared handler
     return await executeGraphQLMutation({
       document: AddProjectCollaboratorDocument,
@@ -23,7 +25,7 @@ export async function addProjectCollaboratorAction({
     });
 
   } catch (error) {
-    logger.error(`[Add Project Collaborator Error]: ${error}`, { error });
+    logger.error(await prepareObjectForLogs({ projectId, email, accessLevel, error }), "Add projectCollaborator");
     return { success: false, errors: ["There was a problem connecting to the server. Please try again."] };
   }
 }