+++ app/[locale]/projects/[projectId]/dmp/[dmpid]/actions/publishPlanAction.ts
@@ -1,7 +1,7 @@
 "use server";
 
 import { executeGraphQLMutation } from "@/utils/server/graphqlServerActionHandler";
-import logger from "@/utils/server/logger";
+import logger, {prepareObjectForLogs} from "@/utils/server/logger";
 import { ActionResponse } from "@/app/types";
 import { PublishPlanDocument } from "@/generated/graphql";
 
@@ -13,6 +13,8 @@ export async function publishPlanAction({
   visibility: string;
 }): Promise<ActionResponse> {
   try {
+    logger.debug(await prepareObjectForLogs({ planId, visibility }), "Publishing plan");
+
     // Execute the mutation using the shared handler
     return await executeGraphQLMutation({
       document: PublishPlanDocument,
@@ -21,7 +23,7 @@ export async function publishPlanAction({
     });
 
   } catch (error) {
-    logger.error(`[Publish Plan Error]: ${error}`, { error });
+    logger.error(await prepareObjectForLogs({ planId, visibility, error }), "Publish plan");
     return { success: false, errors: ["There was a problem connecting to the server. Please try again."] };
   }
 }