+++ app/[locale]/projects/[projectId]/dmp/[dmpid]/actions/updatePlanStatusAction.ts
@@ -1,7 +1,7 @@
 "use server";
 
 import { executeGraphQLMutation } from "@/utils/server/graphqlServerActionHandler";
-import logger from "@/utils/server/logger";
+import logger, {prepareObjectForLogs} from "@/utils/server/logger";
 import { ActionResponse } from "@/app/types";
 import { UpdatePlanStatusDocument } from "@/generated/graphql";
 
@@ -13,6 +13,8 @@ export async function updatePlanStatusAction({
   status: string;
 }): Promise<ActionResponse> {
   try {
+    logger.debug(await prepareObjectForLogs({ planId, status }), "Updating plan status");
+
     // Execute the mutation using the shared handler
     return await executeGraphQLMutation({
       document: UpdatePlanStatusDocument,
@@ -21,7 +23,7 @@ export async function updatePlanStatusAction({
     });
 
   } catch (error) {
-    logger.error(`[Update Plan Status Error]: ${error}`, { error });
+    logger.error(await prepareObjectForLogs({ planId, status, error }), "Update plan status");
     return { success: false, errors: ["There was a problem connecting to the server. Please try again."] };
   }
 }