+++ app/[locale]/projects/[projectId]/dmp/[dmpid]/s/[sid]/q/[qid]/actions/addAnswerCommentAction.ts
@@ -1,7 +1,7 @@
 "use server";
 
 import { executeGraphQLMutation } from "@/utils/server/graphqlServerActionHandler";
-import logger from "@/utils/server/logger";
+import logger, {prepareObjectForLogs} from "@/utils/server/logger";
 import { ActionResponse } from "@/app/types";
 import { AddAnswerCommentDocument } from "@/generated/graphql";
 
@@ -13,6 +13,8 @@ export async function addAnswerCommentAction({
   commentText: string;
 }): Promise<ActionResponse> {
   try {
+    logger.debug(await prepareObjectForLogs({ answerId, commentText }), "Adding answer comment");
+
     // Execute the mutation using the shared handler
     return await executeGraphQLMutation({
       document: AddAnswerCommentDocument,
@@ -20,7 +22,7 @@ export async function addAnswerCommentAction({
       dataPath: "addAnswerComment"
     });
   } catch (error) {
-    logger.error(`[Add answerComment from answer]: ${error}`, { error });
+    logger.error(await prepareObjectForLogs({ answerId, commentText, error }), "Add answerComment for answer");
     return { success: false, errors: ["There was a problem connecting to the server. Please try again."] };
   }
 }