+++ app/[locale]/projects/[projectId]/dmp/[dmpid]/s/[sid]/q/[qid]/actions/updateAnswerCommentAction.ts
@@ -1,7 +1,7 @@
 "use server";
 
 import { executeGraphQLMutation } from "@/utils/server/graphqlServerActionHandler";
-import logger from "@/utils/server/logger";
+import logger, {prepareObjectForLogs} from "@/utils/server/logger";
 import { ActionResponse } from "@/app/types";
 import { UpdateAnswerCommentDocument } from "@/generated/graphql";
 
@@ -15,6 +15,8 @@ export async function updateAnswerCommentAction({
   commentText: string;
 }): Promise<ActionResponse> {
   try {
+    logger.debug(await prepareObjectForLogs({ answerId, answerCommentId, commentText }), "Updating answer comment");
+
     // Execute the mutation using the shared handler
     return await executeGraphQLMutation({
       document: UpdateAnswerCommentDocument,
@@ -22,7 +24,7 @@ export async function updateAnswerCommentAction({
       dataPath: "updateAnswerComment"
     });
   } catch (error) {
-    logger.error(`[Update answerComment from answer]: ${error}`, { error });
+    logger.error(await prepareObjectForLogs({ answerId, answerCommentId, commentText, error }), "Update answerComment for answer");
     return { success: false, errors: ["There was a problem connecting to the server. Please try again."] };
   }
 }