+++ app/[locale]/projects/[projectId]/fundings/[projectFundingId]/edit/actions/removeProjectFundingAction.ts
@@ -1,7 +1,7 @@
 "use server";
 
 import { executeGraphQLMutation } from "@/utils/server/graphqlServerActionHandler";
-import logger from "@/utils/server/logger";
+import logger, {prepareObjectForLogs} from "@/utils/server/logger";
 import { ActionResponse } from "@/app/types";
 import { RemoveProjectFundingDocument } from "@/generated/graphql";
 
@@ -11,6 +11,8 @@ export async function removeProjectFundingAction({
   projectFundingId: number;
 }): Promise<ActionResponse> {
   try {
+    logger.debug(await prepareObjectForLogs({ projectFundingId }), "Remove project funding");
+
     // Execute the mutation using the shared handler
     return await executeGraphQLMutation({
       document: RemoveProjectFundingDocument,
@@ -19,7 +21,7 @@ export async function removeProjectFundingAction({
     });
 
   } catch (error) {
-    logger.error(`[RemoveProjectFunding Error]: ${error}`, { error });
+    logger.error(await prepareObjectForLogs({ projectFundingId, error }), "Remove projectFunding");
     return { success: false, errors: ["There was a problem connecting to the server. Please try again."] };
   }
 }